+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Number of requests handling data_administrators scope (@goreck888)
 - Show pop-up on add to favourites (@maria-j-k)
 - Pre-fill administrator data in catalogues creation form (@maria-j-k)
+- Bundle status changes to draft if any of its offers changes the status from public to other (@maria-j-k)
 - Fix access_type validator in offer (@maria-j-k)